# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/booking": {
            "post": {
                "description": "Create a new booking with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Create a booking",
                "parameters": [
                    {
                        "description": "Booking data",
                        "name": "booking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bookingModule.BookingModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    }
                }
            }
        },
        "/booking/{id}": {
            "get": {
                "description": "Get a booking by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Find a booking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a booking by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Delete a booking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "404": {
                        "description": "Booking not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "Create a new user with the provided information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userModule.UserModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "delete": {
                "description": "Delete a user by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.Answer"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "bookingModule.BookingModel": {
            "type": "object",
            "properties": {
                "delta": {
                    "description": "Длительность бронирования (например, в минутах)",
                    "type": "integer"
                },
                "endTime": {
                    "description": "Время окончания бронирования",
                    "type": "string"
                },
                "id": {
                    "description": "Уникальный идентификатор бронирования",
                    "type": "string"
                },
                "startTime": {
                    "description": "Время начала бронирования",
                    "type": "string"
                },
                "username": {
                    "description": "Имя пользователя, сделавшего бронирование",
                    "type": "string"
                }
            }
        },
        "handlers.Answer": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Данные ответа"
                },
                "status": {
                    "description": "Код состояния HTTP",
                    "type": "integer"
                }
            }
        },
        "userModule.UserModel": {
            "type": "object",
            "properties": {
                "ID": {
                    "description": "Уникальный идентификатор пользователя",
                    "type": "string"
                },
                "createdAt": {
                    "description": "Время создания записи пользователя",
                    "type": "string"
                },
                "password": {
                    "description": "Пароль пользователя (в идеале, храните в хешированном виде)",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "Время последнего обновления записи пользователя",
                    "type": "string"
                },
                "username": {
                    "description": "Имя пользователя (логин)",
                    "type": "string"
                }
            }
        }
    }
}